\name{print.summary.hbrfit}
\alias{print.summary.hbrfit}
\title{print functions for hbrfit}
\description{
These functions print the output in a user-friendly manner using the internal R function \code{print}.
}
\usage{
\method{print}{summary.hbrfit}(x, digits = max(5, .Options$digits - 2), ...)
}
\arguments{
  \item{x}{An object to be printed.}
  \item{digits}{number of digits to display}
  \item{\dots}{additional arguments to be passed to \code{print}}
}
\author{ John Kloke \email{kloke@biostat.wisc.edu}}
